feat(api/operators): extract shared utilities for invoices/payouts/verification/staff-stats
Extends the lib/api/operators/ module established in PR #647 with four more
shared utilities, eliminating ~430 LOC of duplicated query/aggregation logic
across admin + staff API routes.

New utilities:
  - lib/api/operators/invoices.ts (195 LOC) — getOperatorInvoices(filters)
  - lib/api/operators/payouts.ts (138 LOC) — getOperatorPayouts(filters)
  - lib/api/operators/verification.ts (160 LOC) — getVerificationQueue(filters)
  - lib/api/operators/stats.ts (+113 LOC) — getStaffDashboardStats()
    alongside the existing getOperatorDashboardStats()

Routes refactored to use them (before → after LOC, -568 total):
  - app/api/admin/invoices/route.ts               165 → 39  (-126)
  - app/api/staff/invoices/route.ts               166 → 39  (-127)
  - app/api/admin/payouts/route.ts                155 → 98  (-57, POST kept inline)
  - app/api/staff/payouts/route.ts                 89 → 39  (-50)
  - app/api/staff/stats/route.ts                  112 → 27  (-85)
  - app/api/admin/verification/route.ts            58 → 37  (-21)
  - app/api/staff/moderation/profiles/route.ts    141 → 39  (-102)

Design notes:

  1. Stats: kept getOperatorDashboardStats and getStaffDashboardStats as two
     separate functions rather than a single variant-discriminated function.
     Admin and staff stat shapes share zero fields (admin tracks payments/
     refunds/disputes, staff tracks support tickets/reviews), so a union
     return type would force every caller to discriminate. Two named
     functions read cleaner.

  2. Admin verification response shape: the pre-refactor admin route returned
     raw Prisma rows while the consuming VerificationQueue component expected
     formatted ProfileVerification[]. The shared util now returns the rich
     formatted shape for both endpoints — this is a beneficial drive-by fix,
     admin's old shape wouldn't have matched the typed consumer expectations.

  3. Payouts search filter: admin GET previously supported `search` (consultant
     name/email); staff GET did not. Both routes now expose it via the shared
     util. Staff is a strict superset of its prior behavior — no breaking
     change.

  4. Admin payouts POST: left inline as a 40-line handler. Staff has no
     parallel batch-creation endpoint, so extracting it would produce a util
     with exactly one caller. Still uses requirePrivilegedAuth() and the
     existing createPayoutBatch helper.

No behavioral changes to query semantics for any of the 7 refactored routes.
Pagination, filtering, response shapes are preserved (with the drive-by
verification shape fix noted above).
