+++ app/api/staff/invoices/route.ts
@@ -1,13 +1,16 @@
 /**
  * Staff Invoices API
- * Staff access to invoice viewing
+ * Staff access to invoice viewing.
+ *
+ * Thin shell — query/response logic lives in `lib/api/operators/invoices.ts`
+ * and is shared with `/api/admin/invoices`.
  */
 
 import { NextRequest, NextResponse } from "next/server";
-import prisma from "@/lib/prisma";
-import { PaymentStatus, Prisma } from "@prisma/client";
-
+import { PaymentStatus } from "@prisma/client";
 import { requirePrivilegedAuth } from "@/lib/auth-helpers";
+import { getOperatorInvoices } from "@/lib/api/operators";
+
 /**
  * GET /api/staff/invoices
  * Get all invoices with optional filters (staff access)
@@ -16,146 +19,16 @@ export async function GET(req: NextRequest) {
   try {
     const auth = await requirePrivilegedAuth();
     if (auth.error) return auth.error;
-    const session = auth.session;
 
-    // Parse query parameters
     const { searchParams } = new URL(req.url);
-    const status = searchParams.get("status") as PaymentStatus | null;
-    const search = searchParams.get("search");
-    const limit = parseInt(searchParams.get("limit") || "20");
-    const offset = parseInt(searchParams.get("offset") || "0");
-
-    // Build where clause
-    const where: Prisma.PaymentWhereInput = {};
-    if (status) {
-      where.paymentStatus = status;
-    }
-    if (search) {
-      where.OR = [
-        { paymentIntent: { contains: search, mode: "insensitive" } },
-        {
-          user: {
-            OR: [
-              { name: { contains: search, mode: "insensitive" } },
-              { email: { contains: search, mode: "insensitive" } },
-            ],
-          },
-        },
-      ];
-    }
-
-    // Get invoices (payments with succeeded status are considered invoices)
-    const [payments, total] = await Promise.all([
-      prisma.payment.findMany({
-        where: {
-          ...where,
-          paymentStatus: status || "SUCCEEDED",
-        },
-        include: {
-          user: {
-            select: { name: true, email: true },
-          },
-          appointment: {
-            include: {
-              consultation: {
-                include: {
-                  consultationPlan: {
-                    include: {
-                      consultantProfile: {
-                        include: { user: { select: { name: true } } },
-                      },
-                    },
-                  },
-                },
-              },
-              subscription: {
-                include: {
-                  subscriptionPlan: {
-                    include: {
-                      consultantProfile: {
-                        include: { user: { select: { name: true } } },
-                      },
-                    },
-                  },
-                },
-              },
-              webinar: {
-                include: {
-                  webinarPlan: {
-                    include: {
-                      consultantProfile: {
-                        include: { user: { select: { name: true } } },
-                      },
-                    },
-                  },
-                },
-              },
-              class: {
-                include: {
-                  classPlan: {
-                    include: {
-                      consultantProfile: {
-                        include: { user: { select: { name: true } } },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-        orderBy: { createdAt: "desc" },
-        take: limit,
-        skip: offset,
-      }),
-      prisma.payment.count({
-        where: {
-          ...where,
-          paymentStatus: status || "SUCCEEDED",
-        },
-      }),
-    ]);
-
-    // Helper to extract consultant name from different appointment types
-    const getConsultantName = (
-      appointment: (typeof payments)[0]["appointment"],
-    ) => {
-      if (!appointment) return undefined;
-      return (
-        appointment.consultation?.consultationPlan?.consultantProfile?.user
-          ?.name ||
-        appointment.subscription?.subscriptionPlan?.consultantProfile?.user
-          ?.name ||
-        appointment.webinar?.webinarPlan?.consultantProfile?.user?.name ||
-        appointment.class?.classPlan?.consultantProfile?.user?.name
-      );
-    };
-
-    // Format as invoices
-    const invoices = payments.map((p) => ({
-      id: p.id,
-      invoiceNumber: `INV-${p.id.slice(-8).toUpperCase()}`,
-      paymentIntent: p.paymentIntent,
-      amount: p.amount,
-      currency: p.currency,
-      status: p.paymentStatus,
-      gateway: p.paymentGateway,
-      userName: p.user?.name || "Unknown",
-      userEmail: p.user?.email || "",
-      appointmentType: p.appointment?.appointmentType,
-      consultantName: getConsultantName(p.appointment),
-      createdAt: p.createdAt,
-    }));
-
-    return NextResponse.json({
-      invoices,
-      pagination: {
-        total,
-        limit,
-        offset,
-        hasMore: offset + limit < total,
-      },
+    const result = await getOperatorInvoices({
+      status: searchParams.get("status") as PaymentStatus | null,
+      search: searchParams.get("search"),
+      limit: parseInt(searchParams.get("limit") || "20"),
+      offset: parseInt(searchParams.get("offset") || "0"),
     });
+
+    return NextResponse.json(result);
   } catch (error) {
     console.error("Error fetching invoices:", error);
     return NextResponse.json(
